+++ src/fwrite.c
@@ -619,8 +619,8 @@ void fwriteMain(fwriteMainArgs args)
       DTPRINT(_("... "));
       for (int j=args.ncol-10; j<args.ncol; j++) DTPRINT(_("%d "), args.whichFun[j]);
     }
-    DTPRINT(_("\nargs.doRowNames=%d args.rowNames=%d doQuote=%d args.nrow=%"PRIu64" args.ncol=%d eolLen=%d\n"),
-          args.doRowNames, args.rowNames, doQuote, (long long)args.nrow, args.ncol, eolLen);
+    DTPRINT(_("\nargs.doRowNames=%d args.rowNames=%d doQuote=%d args.nrow=%"PRId64" args.ncol=%d eolLen=%d\n"),
+          args.doRowNames, args.rowNames, doQuote, (int64_t)args.nrow, args.ncol, eolLen);
   }
 
   // Calculate upper bound for line length. Numbers use a fixed maximum (e.g. 12 for integer) while strings find the longest
@@ -660,7 +660,7 @@ void fwriteMain(fwriteMainArgs args)
     maxLineLen += width*2;  // *2 in case the longest string is all quotes and they all need to be escaped
   }
 
-  if (verbose) DTPRINT(_("maxLineLen=%"PRIu64". Found in %.3fs\n"), maxLineLen, 1.0*(wallclock()-t0));
+  if (verbose) DTPRINT(_("maxLineLen=%"PRIu64". Found in %.3fs\n"), (uint64_t)maxLineLen, 1.0*(wallclock()-t0));
 
   int f=0;
   if (*args.filename=='\0') {
@@ -778,8 +778,8 @@ void fwriteMain(fwriteMainArgs args)
   int nth = args.nth;
   if (numBatches < nth) nth = numBatches;
   if (verbose) {
-    DTPRINT(_("Writing %"PRIu64" rows in %d batches of %d rows (each buffer size %dMB, showProgress=%d, nth=%d)\n"),
-            (long long)args.nrow, numBatches, rowsPerBatch, args.buffMB, args.showProgress, nth);
+    DTPRINT(_("Writing %"PRId64" rows in %d batches of %d rows (each buffer size %dMB, showProgress=%d, nth=%d)\n"),
+            (int64_t)args.nrow, numBatches, rowsPerBatch, args.buffMB, args.showProgress, nth);
   }
   t0 = wallclock();
